ui: introduce keyboard scroll flag in Virtual Console
Introduce in Virtual Console Settings a flag to enable/disable
standard keyboard management on scroll area.
The flag, when enabled (default), guarantees current behaviour:
arrow keys and PgUp/PgDown keys are used to scroll the console
and cannot be used as key combinations for controls.
When disabled, the scroll area does not consume key events
anymore, leaving all the keys available as key combinations
on the controls.